requirements.txt
PyMySQL==0.6.3
Flask-SQLAlchemy==2.0
SQLAlchemy==0.9.8